Fix file structure issues
