--- src/Components/App/App.js
@@ -1,11 +1,0 @@
-
-import './App.css';
-
-const App = () => {
-
-}
-
-return (
-
-)
-export default App;
