+++ src/Components/Home/Home.js
@@ -1,0 +1,2 @@
+import { useState } from 'react'
+import '../../../Home.css'
